+++ src/main/java/com/cliffc/aa/type/TypeRPC.java
@@ -63,7 +63,7 @@ public class TypeRPC extends Type<TypeRPC> {
     return make(_rpcs.meet( tf._rpcs ));
   }
   
-  public int rpc() { return _rpcs.getbit(); }
+  public long rpc() { return _rpcs.getbit(); }
   public boolean test(int rpc) { return _rpcs.test(rpc); }
   @Override public boolean above_center() { return _rpcs.above_center(); }
   @Override public boolean may_be_con()   { return _rpcs.is_con() || _rpcs.above_center(); }
